Compiling

+++ MOA/Mapping.cpp
@@ -3,9 +3,11 @@
 #include <boost/property_tree/ptree.hpp>
 #include <boost/property_tree/json_parser.hpp>
 
+
+
 using boost::property_tree::ptree;
 using boost::property_tree::read_json;
- 
+
 using namespace std;
 
 namespace moa {
@@ -17,11 +19,11 @@ constexpr tuple<int, int> get2dCoordinates(ptree::value_type const &entry) {
   return make_tuple(x, y);
 }
 
-constexpr tuple<int, int, int>
+constexpr tuple<double, double, double>
 get3dCoordinates(ptree::value_type const &entry) {
-  const int x = entry.second.get<int>("3d-X");
-  const int y = entry.second.get<int>("3d-Y");
-  const int z = entry.second.get<int>("3d-Z");
+  const double x = entry.second.get<double>("3d-X");
+  const double y = entry.second.get<double>("3d-Y");
+  const double z = entry.second.get<double>("3d-Z");
   return make_tuple(x, y, z);
 }
